+++ docs/conf.py
@@ -27,7 +27,7 @@ author = 'David Seddon'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '1.0b4'
+release = '1.0b5'
 
 
 # -- General configuration ---------------------------------------------------