+++ src/importlinter/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0b4"
+__version__ = "1.0b5"
 
 from .application import output  # noqa
 from .domain import fields  # noqa